PR feedback. Revert support for immutable schemas.
