+++ realm/realm-library/src/main/java/io/realm/RealmObjectSchema.java
@@ -402,29 +402,13 @@ public abstract class RealmObjectSchema {
      * Runs a transformation function on each RealmObject instance of the current class. The object will be represented
      * as a {@link DynamicRealmObject}.
      * <p>
-     * There is no guarantees in which order the objects are re
+     * There is no guarantees in which order the objects are returned.
      *
      * @param function transformation function.
      * @return this schema.
      * @throws UnsupportedOperationException if this {@link RealmObjectSchema} is immutable.
      */
-    public RealmObjectSchema transform(Function function) {
-        //noinspection ConstantConditions
-        if (function != null) {
-            // Users might delete object being transformed or accidentally delete other objects
-            // in the same table. E.g. cascading deletes if it is referenced by an object being deleted.
-            OsResults results = OsResults.createFromTable(realm.sharedRealm, table).createSnapshot();
-            long size = results.size();
-            for (long i = 0; i < size; i++) {
-                DynamicRealmObject obj = new DynamicRealmObject(realm, new CheckedRow(results.getUncheckedRow((int) i)));
-                if (obj.isValid()) {
-                    function.apply(obj);
-                }
-            }
-        }
-
-        return this;
-    }
+    public abstract RealmObjectSchema transform(Function function);
 
     /**
      * Returns the type used by the underlying storage engine to represent this field.
